requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-rest-captcha==0.1.0
djangorestframework==3.11.1
olefile==0.46
Pillow==4.3.0
pytz==2020.1
sqlparse==0.3.1
